Calibration: Changed constants from integerst to double floating point numbers in the perspective calibration application. #147

+++ user/applications/calibration/perspective/Checkerboard.hpp
@@ -40,9 +40,9 @@ namespace rtx {
       unsigned int filled;
 
       // Perspective constants
-      unsigned int A;
-      unsigned int B;
-      unsigned int C;
+      double A;
+      double B;
+      double C;
 
   };
